pyproject.toml
[tool.poetry]
name = "hospitalsystem"
version = "0.1.0"
description = ""
authors = ["Andrew Demchuk"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.1"
django-grappelli = "4.0.1"
environs = "11.0.0"
psycopg2-binary = "^2.9.9"
gunicorn="23.0.0"
django-cors-headers="4.4.0"
djangorestframework="3.15.2"
django-allauth="64.2.1"
Pillow="10.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
